% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{apply_scaling_enhanced}
\alias{apply_scaling_enhanced}
\title{Apply Enhanced Scaling Methods}
\usage{
apply_scaling_enhanced(matrix_data, scale_method, verbose = FALSE)
}
\arguments{
\item{matrix_data}{Numeric matrix to scale}

\item{scale_method}{Scaling method: "variable_0_10", "robust", "row", "column", "none"}

\item{verbose}{Whether to print scaling information}
}
\value{
Scaled matrix
}
\description{
Applies various scaling methods to matrix data for heatmap visualization
}
\keyword{internal}
