% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{print_detailed_summary}
\alias{print_detailed_summary}
\title{Print Detailed PCA Variable Summary}
\usage{
print_detailed_summary(
  top_vars,
  pc_x_top,
  pc_y_top,
  high_both,
  pc_x,
  pc_y,
  top_n,
  min_loading_threshold
)
}
\arguments{
\item{top_vars}{Data frame of top variables by combined importance}

\item{pc_x_top}{Data frame of top variables for first PC}

\item{pc_y_top}{Data frame of top variables for second PC}

\item{high_both}{Data frame of variables important in both PCs}

\item{pc_x}{Name of first principal component}

\item{pc_y}{Name of second principal component}

\item{top_n}{Number of top variables to display}

\item{min_loading_threshold}{Minimum loading threshold}
}
\value{
NULL (prints to console)
}
\description{
Prints formatted summary of PCA variable importance analysis
}
\keyword{internal}
