% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{handle_missing_values}
\alias{handle_missing_values}
\title{Standardize Timepoint Names}
\usage{
handle_missing_values(data, value_column, method, verbose)
}
\arguments{
\item{timepoints}{Character vector of timepoint names}

\item{naming_scheme}{Character. Standardization scheme to apply}

\item{...}{Additional parameters}
}
\value{
Character vector with standardized timepoint names
}
\description{
Template function for standardizing timepoint naming across experiments
}
\examples{
\dontrun{
# Standardize timepoint names
std_timepoints <- standardize_timepoint_names(c("1h", "1hr", "60min"))
}

}
