% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{setup_color_scheme}
\alias{setup_color_scheme}
\title{Setup Color Scheme}
\usage{
setup_color_scheme(color_scheme, custom_colors)
}
\arguments{
\item{color_scheme}{Name of color scheme to use}

\item{custom_colors}{Custom color list (optional)}
}
\value{
List of colors for plotting
}
\description{
Sets up color schemes for plotting functions
}
